Don't re-use taskIds for resident tasks
This introduces the notion of launch attempts for resident tasks. With this change, every task that is launched on an existing reservation will use a new taskId that has an incremented attempt counter. Matching reservations and volumes is done against the base reservationId, which is basically the taskId without separator and attempt count. There is no migration needed, but consumers of taskIds need to be aware that newly launched resident tasks will now have an attempt appended to the ID, e.g. `app.b6ff5fa5-7714-11e7-a55c-5ecf1c4671f6.#42`, which denotes attempt number 42.

There are additional changes in this set that consolidate the usage of task vs instance, e.g. in Tasklabels.

JIRA: MARATHON_EE-1603